+++ nmclicktests/environment.js
@@ -1,4 +1,5 @@
 'use strict';
+const winston = require('winston');
 const child_process = require('child_process');
 const Bluebird = require('bluebird');
 const Nightmare = require('nightmare');
@@ -13,14 +14,14 @@ module.exports = (config) => new Environment(config);
 
 Nightmare.action('ug', {
   'log': function(message, done) {
-    console.log(`>>> ${message}`);
+    winston.info(`>>> ${message}`);
     done();
   },
   'commit': function(commitMessage, done) {
     this.wait('.files .file .btn-default')
       .insert('.staging input.form-control', commitMessage)
       .wait(100)
-      .click('.commit-btn')
+      .ug.click('.commit-btn')
       .ug.waitForElementNotVisible('.files .file .btn-default')
       .wait(1000)
       .then(done.bind(null, null), done);
@@ -34,14 +35,14 @@ Nightmare.action('ug', {
   },
   'amendCommit': function(done) {
     this.ug.click('.amend-link')
-      .click('.commit-btn')
+      .ug.click('.commit-btn')
       .ug.waitForElementNotVisible('.files .file .btn-default')
       .wait(1000)
       .then(done.bind(null, null), done);
   },
   'emptyCommit': function(done) {
     this.ug.click('.empty-commit-link')
-      .click('.commit-btn')
+      .ug.click('.commit-btn')
       .ug.waitForElementNotVisible('.files .file .btn-default')
       .wait(1000)
       .then(done.bind(null, null), done);
@@ -91,12 +92,12 @@ Nightmare.action('ug', {
       .then(done.bind(null, null), done);
   },
   'createTempFolder': function(done) {
-    console.log('Creating temp folder');
+    winston.info('Creating temp folder');
     this.ug.backgroundAction('POST', `${rootUrl}/api/testing/createtempdir`, undefined)
       .then(done.bind(null, null), done);
   },
   'createFolder': function(dir, done) {
-    console.log(`Create folder: ${dir}`);
+    winston.info(`Create folder: ${dir}`);
     this.ug.backgroundAction('POST', `${rootUrl}/api/createdir`, { dir: dir })
       .then(done.bind(null, null), done);
   },
@@ -139,6 +140,7 @@ Nightmare.action('ug', {
     this.ug.click('.current ~ .newRef button.showBranchingForm')
       .insert('.newRef.editing input', name)
       .wait(100)
+      // cannot use .ug.click as wait op will defocus and doms will disappear
       .click(`.newRef ${type === 'branch' ? '.btn-primary' : '.btn-default'}`)
       .wait(`.ref.${type}[data-ta-name="${name}"]`)
       .wait(300)
@@ -187,10 +189,10 @@ class Environment {
     // init
     this.nm.viewport(this.config.viewWidth, this.config.viewHeight);
     this.nm.on('console', (type, msg1, msg2) => {
-      console.log(`[ui ${type}] ${(new Date()).toISOString()}  - ${msg1} ${JSON.stringify(msg2)}`);
+      winston.info(`[ui ${type}] ${(new Date()).toISOString()}  - ${msg1} ${JSON.stringify(msg2)}`);
 
       if (type === 'error' && !this.shuttinDown) {
-        console.log('ERROR DETECTED!');
+        winston.info('ERROR DETECTED!');
       }
     })
   }
@@ -232,7 +234,7 @@ class Environment {
     return this.getPort()
       .then(() => this.startServer())
       .then(() => this.ensureStarted())
-      .catch((err) => { console.log(err); throw new Error("Cannot confirm ungit start!!", err); })
+      .catch((err) => { winston.error(err); throw new Error("Cannot confirm ungit start!!", err); })
   }
 
   createRepos(testRepoPaths, config) {
@@ -277,7 +279,7 @@ class Environment {
   }
 
   startServer() {
-    console.log('Starting ungit server...', this.config.serverStartupOptions);
+    winston.info('Starting ungit server...', this.config.serverStartupOptions);
 
     this.hasStarted = false;
     const options = ['bin/ungit',
@@ -298,31 +300,31 @@ class Environment {
     const ungitServer = child_process.spawn('node', options);
     ungitServer.stdout.on('data', (stdout) => {
       const stdoutStr = stdout.toString();
-      if (this.config.showServerOutput) console.log(prependLines('[server] ', stdoutStr));
+      if (this.config.showServerOutput) winston.verbose(prependLines('[server] ', stdoutStr));
 
       if (stdoutStr.indexOf('Ungit server already running') >= 0) {
-        console.log('server-already-running');
+        winston.info('server-already-running');
       }
 
       if (stdoutStr.indexOf('## Ungit started ##') >= 0) {
         if (this.hasStarted) {
-          console.log('Ungit started twice, probably crashed.');
+          winston.info('Ungit started twice, probably crashed.');
         } else {
           this.hasStarted = true;
-          console.log('Ungit server started.');
+          winston.info('Ungit server started.');
         }
       }
     });
     ungitServer.stderr.on("data", (stderr) => {
       const stderrStr = stderr.toString();
-      console.error(prependLines('[server ERROR] ', stderrStr));
+      winston.error(prependLines('[server ERROR] ', stderrStr));
       if (stderrStr.indexOf("EADDRINUSE") > -1) {
-        console.log("retrying with different port");
+        winston.info("retrying with different port");
         ungitServer.kill('SIGINT');
         this.getPort().then(() => this.startServer());
       }
     });
-    ungitServer.on('exit', () => console.log('UNGIT SERVER EXITED'));
+    ungitServer.on('exit', () => winston.info('UNGIT SERVER EXITED'));
     return Bluebird.resolve();
   }
 }